Renamed some C-API functions related to object management for a more intuitive grouping. Reconciled related tests and examples.
