+++ test/src/unit-capi-object_mgmt.cc
@@ -1,5 +1,5 @@
 /**
- * @file   unit-capi-walk.cc
+ * @file   unit-capi-object_mgmt.cc
  *
  * @section LICENSE
  *
@@ -27,58 +27,66 @@
  *
  * @section DESCRIPTION
  *
- * Tests for the C API walk code.
+ * Tests for the C API object management code.
  */
 
 #include "catch.hpp"
-#include "constants.h"
+
 #ifdef _WIN32
 #include "win_filesystem.h"
 #else
 #include "posix_filesystem.h"
 #endif
+
 #include "tiledb.h"
-#include "uri.h"
 
-#include <iostream>
-
-struct WalkFx {
+struct ObjectMgmtFx {
 #ifdef HAVE_HDFS
   const std::string HDFS_TEMP_DIR = "hdfs:///tiledb_test/";
   const std::string HDFS_FULL_TEMP_DIR = "hdfs://localhost:9000/tiledb_test/";
 #endif
 #ifdef HAVE_S3
-  const tiledb::URI S3_BUCKET = tiledb::URI("s3://tiledb/");
+  const tiledb::URI S3_BUCKET = tiledb::URI("s3://tiledb");
   const std::string S3_TEMP_DIR = "s3://tiledb/tiledb_test/";
 #endif
-
 #ifdef _WIN32
-  const std::string FILE_TEMP_DIR = "tiledb_test\\";
-  const std::string FILE_FULL_TEMP_DIR =
+  const std::string FILE_URI_PREFIX = "";
+  const std::string FILE_TEMP_DIR =
       tiledb::win::current_dir() + "\\tiledb_test\\";
+  const std::string FILE_FULL_TEMP_DIR = FILE_TEMP_DIR;
+  const std::string GROUP = "group\\";
+  const std::string ARRAY = "array\\";
 #else
-  const std::string FILE_TEMP_DIR = "tiledb_test/";
-  const std::string FILE_FULL_TEMP_DIR =
-      std::string("file://") + tiledb::posix::current_dir() + "/tiledb_test/";
+  const std::string FILE_URI_PREFIX = "file://";
+  const std::string FILE_TEMP_DIR =
+      tiledb::posix::current_dir() + "/tiledb_test/";
+  const std::string FILE_FULL_TEMP_DIR = std::string("file://") + FILE_TEMP_DIR;
+  const std::string GROUP = "group/";
+  const std::string ARRAY = "array/";
 #endif
 
-  // TileDB context and VFS
-  tiledb_ctx_t *ctx_;
-  tiledb_vfs_t *vfs_;
+  // TileDB context
+  tiledb_ctx_t* ctx_;
+  tiledb_vfs_t* vfs_;
 
   // Functions
-  WalkFx();
-  ~WalkFx();
-  void remove_temp_dir(const std::string &path);
-  void create_hierarchy(const std::string &path);
-  std::string get_golden_walk(const std::string &path);
-  std::string get_golden_ls(const std::string &path);
-  static int write_path(const char *path, tiledb_object_t type, void *data);
+  ObjectMgmtFx();
+  ~ObjectMgmtFx();
+  void check_object_type(const std::string& path);
+  void check_delete(const std::string& path);
+  void check_move(const std::string& path);
+  void create_array(const std::string& path);
+  void create_temp_dir(const std::string& path);
+  void remove_temp_dir(const std::string& path);
+  void create_hierarchy(const std::string& path);
+  std::string get_golden_walk(const std::string& path);
+  std::string get_golden_ls(const std::string& path);
+  static int write_path(const char* path, tiledb_object_t type, void* data);
 };
 
-WalkFx::WalkFx() {
+ObjectMgmtFx::ObjectMgmtFx() {
   // Create TileDB context
-  tiledb_config_t *config = nullptr;
+  tiledb_config_t* config = nullptr;
   REQUIRE(tiledb_config_create(&config) == TILEDB_OK);
 #ifdef HAVE_S3
   REQUIRE(
@@ -103,16 +111,154 @@ WalkFx::WalkFx() {
 #endif
 }
 
-WalkFx::~WalkFx() {
+ObjectMgmtFx::~ObjectMgmtFx() {
   CHECK(tiledb_vfs_free(ctx_, vfs_) == TILEDB_OK);
   CHECK(tiledb_ctx_free(ctx_) == TILEDB_OK);
 }
 
-void WalkFx::remove_temp_dir(const std::string &path) {
+void ObjectMgmtFx::create_temp_dir(const std::string& path) {
+  remove_temp_dir(path);
+  REQUIRE(tiledb_vfs_create_dir(ctx_, vfs_, path.c_str()) == TILEDB_OK);
+}
+
+void ObjectMgmtFx::remove_temp_dir(const std::string& path) {
   int is_dir = 0;
   REQUIRE(tiledb_vfs_is_dir(ctx_, vfs_, path.c_str(), &is_dir) == TILEDB_OK);
   if (is_dir)
     REQUIRE(tiledb_vfs_remove_dir(ctx_, vfs_, path.c_str()) == TILEDB_OK);
+}
+
+void ObjectMgmtFx::create_array(const std::string& path) {
+  tiledb_attribute_t* a1;
+  tiledb_attribute_create(ctx_, &a1, "a1", TILEDB_FLOAT32);
+
+  // Domain and tile extents
+  int64_t dim_domain[2] = {1, 1};
+  int64_t tile_extents[] = {1};
+
+  // Create dimension
+  tiledb_dimension_t* d1;
+  tiledb_dimension_create(
+      ctx_, &d1, "d1", TILEDB_INT64, &dim_domain[0], &tile_extents[0]);
+
+  // Domain
+  tiledb_domain_t* domain;
+  tiledb_domain_create(ctx_, &domain);
+  tiledb_domain_add_dimension(ctx_, domain, d1);
+
+  // Create array schema
+  tiledb_array_schema_t* array_schema;
+  tiledb_array_schema_create(ctx_, &array_schema);
+  tiledb_array_schema_set_array_type(ctx_, array_schema, TILEDB_DENSE);
+  tiledb_array_schema_set_domain(ctx_, array_schema, domain);
+  tiledb_array_schema_add_attribute(ctx_, array_schema, a1);
+
+  // Check array schema
+  REQUIRE(tiledb_array_schema_check(ctx_, array_schema) == TILEDB_OK);
+
+  // Create array
+  REQUIRE(tiledb_array_create(ctx_, path.c_str(), array_schema) == TILEDB_OK);
+  tiledb_dimension_free(ctx_, d1);
+}
+
+void ObjectMgmtFx::check_object_type(const std::string& path) {
+  std::string group, array;
+  tiledb_object_t type;
+
+  // Check group
+  group = path + "group/";
+  REQUIRE(tiledb_group_create(ctx_, group.c_str()) == TILEDB_OK);
+  REQUIRE(tiledb_object_type(ctx_, group.c_str(), &type) == TILEDB_OK);
+  CHECK(type == TILEDB_GROUP);
+
+  // Check invalid
+  array = group + "array/";
+  REQUIRE(tiledb_object_type(ctx_, array.c_str(), &type) == TILEDB_OK);
+  CHECK(type == TILEDB_INVALID);
+
+  // Check array
+  create_array(array);
+  REQUIRE(tiledb_object_type(ctx_, array.c_str(), &type) == TILEDB_OK);
+  CHECK(type == TILEDB_ARRAY);
+}
+
+void ObjectMgmtFx::check_delete(const std::string& path) {
+  std::string group, array, invalid;
+  tiledb_object_t type;
+
+  // Check simple delete
+  group = path + "group/";
+  CHECK(tiledb_object_remove(ctx_, group.c_str()) == TILEDB_OK);
+
+  // Check invalid delete
+  invalid = group + "foo";
+  CHECK(tiledb_object_remove(ctx_, invalid.c_str()) == TILEDB_ERR);
+
+  // Check recursive delete
+  REQUIRE(tiledb_group_create(ctx_, group.c_str()) == TILEDB_OK);
+  REQUIRE(tiledb_group_create(ctx_, (group + "l1").c_str()) == TILEDB_OK);
+  REQUIRE(tiledb_group_create(ctx_, (group + "l1/l2").c_str()) == TILEDB_OK);
+  REQUIRE(tiledb_group_create(ctx_, (group + "l1/l2/l3").c_str()) == TILEDB_OK);
+  REQUIRE(tiledb_object_type(ctx_, (group + "l1").c_str(), &type) == TILEDB_OK);
+  CHECK(type == TILEDB_GROUP);
+  REQUIRE(
+      tiledb_object_type(ctx_, (group + "l1/l2").c_str(), &type) == TILEDB_OK);
+  CHECK(type == TILEDB_GROUP);
+  REQUIRE(
+      tiledb_object_type(ctx_, (group + "l1/l2/l3").c_str(), &type) ==
+      TILEDB_OK);
+  CHECK(type == TILEDB_GROUP);
+  REQUIRE(tiledb_object_remove(ctx_, (group + "l1").c_str()) == TILEDB_OK);
+  REQUIRE(
+      tiledb_object_type(ctx_, (group + "l1/l2/l3").c_str(), &type) ==
+      TILEDB_OK);
+  CHECK(type == TILEDB_INVALID);
+  REQUIRE(
+      tiledb_object_type(ctx_, (group + "l1/l2").c_str(), &type) == TILEDB_OK);
+  CHECK(type == TILEDB_INVALID);
+  REQUIRE(tiledb_object_type(ctx_, (group + "l1").c_str(), &type) == TILEDB_OK);
+  CHECK(type == TILEDB_INVALID);
+}
+
+void ObjectMgmtFx::check_move(const std::string& path) {
+  // Move group
+  auto group = path + "group/";
+  auto old1 = group + "old1";
+  auto old2 = group + "old2";
+  auto new1 = group + "new1";
+  auto new2 = group + "new2";
+  REQUIRE(tiledb_group_create(ctx_, old1.c_str()) == TILEDB_OK);
+  REQUIRE(tiledb_group_create(ctx_, old2.c_str()) == TILEDB_OK);
+  CHECK(
+      tiledb_object_move(ctx_, old1.c_str(), new1.c_str(), false) == TILEDB_OK);
+
+  tiledb_object_t type;
+  CHECK(tiledb_object_type(ctx_, new1.c_str(), &type) == TILEDB_OK);
+  CHECK(type == TILEDB_GROUP);
+
+  // Check error on name conflict
+  CHECK(
+      tiledb_object_move(ctx_, new1.c_str(), old2.c_str(), false) ==
+      TILEDB_ERR);
+
+  // Check force move works on name conflict
+  CHECK(
+      tiledb_object_move(ctx_, new1.c_str(), old2.c_str(), true) == TILEDB_OK);
+
+  // Check move array
+  auto array = group + ARRAY;
+  auto array2 = group + "new_array";
+  create_array(array);
+  CHECK(
+      tiledb_object_move(ctx_, array.c_str(), array2.c_str(), false) ==
+      TILEDB_OK);
+
+  // Check error on invalid path
+  auto inv1 = path + "invalid_path";
+  auto inv2 = path + "new_invalid_path";
+  CHECK(
+      tiledb_object_move(ctx_, inv1.c_str(), inv2.c_str(), false) ==
+      TILEDB_ERR);
 }
 
 /**
@@ -134,7 +280,7 @@ void WalkFx::remove_temp_dir(const std::string &path) {
  *            |_ array_D
  *                  |_ __array_schema.tdb
  */
-void WalkFx::create_hierarchy(const std::string &path) {
+void ObjectMgmtFx::create_hierarchy(const std::string& path) {
   int rc = tiledb_vfs_create_dir(ctx_, vfs_, path.c_str());
   REQUIRE(rc == TILEDB_OK);
   rc = tiledb_vfs_create_dir(ctx_, vfs_, (path + "dense_arrays").c_str());
@@ -178,7 +324,7 @@ void WalkFx::create_hierarchy(const std::string &path) {
   REQUIRE(rc == TILEDB_OK);
 }
 
-std::string WalkFx::get_golden_walk(const std::string &path) {
+std::string ObjectMgmtFx::get_golden_walk(const std::string& path) {
   std::string golden;
 
   // Preorder traversal
@@ -202,7 +348,7 @@ std::string WalkFx::get_golden_walk(const std::string &path) {
   return golden;
 }
 
-std::string WalkFx::get_golden_ls(const std::string &path) {
+std::string ObjectMgmtFx::get_golden_ls(const std::string& path) {
   std::string golden;
 
   golden += path + "dense_arrays GROUP\n";
@@ -211,9 +357,10 @@ std::string WalkFx::get_golden_ls(const std::string &path) {
   return golden;
 }
 
-int WalkFx::write_path(const char *path, tiledb_object_t type, void *data) {
+int ObjectMgmtFx::write_path(
+    const char* path, tiledb_object_t type, void* data) {
   // Cast data to string
-  auto *str = static_cast<std::string *>(data);
+  auto* str = static_cast<std::string*>(data);
 
   // Simply print the path and type
   (*str) += (std::string(path) + " ");
@@ -236,8 +383,42 @@ int WalkFx::write_path(const char *path, tiledb_object_t type, void *data) {
   return 1;
 }
 
-TEST_CASE_METHOD(WalkFx, "C API: Test walk", "[capi], [walk]") {
-  std::string golden_walk, golden_ls, walk_str, ls_str;
+TEST_CASE_METHOD(
+    ObjectMgmtFx,
+    "C API: Test object management methods: object_type, delete, move",
+    "[capi], [object], [delete], [move], [object_type]") {
+  // File
+  create_temp_dir(FILE_URI_PREFIX + FILE_TEMP_DIR);
+  check_object_type(FILE_URI_PREFIX + FILE_TEMP_DIR);
+  check_delete(FILE_URI_PREFIX + FILE_TEMP_DIR);
+  check_move(FILE_URI_PREFIX + FILE_TEMP_DIR);
+  remove_temp_dir(FILE_URI_PREFIX + FILE_TEMP_DIR);
+
+// S3
+#ifdef HAVE_S3
+  create_temp_dir(S3_TEMP_DIR);
+  check_object_type(S3_TEMP_DIR);
+  check_delete(S3_TEMP_DIR);
+  check_move(S3_TEMP_DIR);
+  remove_temp_dir(S3_TEMP_DIR);
+#endif
+
+// HDFS
+#ifdef HAVE_HDFS
+  create_temp_dir(HDFS_TEMP_DIR);
+  check_object_type(HDFS_TEMP_DIR);
+  check_delete(HDFS_TEMP_DIR);
+  check_move(HDFS_TEMP_DIR);
+  remove_temp_dir(HDFS_TEMP_DIR);
+#endif
+}
+
+TEST_CASE_METHOD(
+    ObjectMgmtFx,
+    "C API: Test object management methods: walk, ls",
+    "[capi], [object], [walk], [ls]") {
+  std::string golden_walk, golden_ls;
+  std::string walk_str, ls_str;
   int rc;
 
   // File
@@ -252,10 +433,11 @@ TEST_CASE_METHOD(WalkFx, "C API: Test walk", "[capi], [walk]") {
   golden_ls = get_golden_ls(FILE_FULL_TEMP_DIR);
 #endif
   walk_str.clear();
-  rc = tiledb_walk(
+  ls_str.clear();
+  rc = tiledb_object_walk(
       ctx_, FILE_FULL_TEMP_DIR.c_str(), TILEDB_PREORDER, write_path, &walk_str);
   CHECK(rc == TILEDB_OK);
-  rc = tiledb_walk(
+  rc = tiledb_object_walk(
       ctx_,
       FILE_FULL_TEMP_DIR.c_str(),
       TILEDB_POSTORDER,
@@ -275,10 +457,10 @@ TEST_CASE_METHOD(WalkFx, "C API: Test walk", "[capi], [walk]") {
   golden_ls = get_golden_ls(S3_TEMP_DIR);
   walk_str.clear();
   ls_str.clear();
-  rc = tiledb_walk(
+  rc = tiledb_object_walk(
       ctx_, S3_TEMP_DIR.c_str(), TILEDB_PREORDER, write_path, &walk_str);
   CHECK(rc == TILEDB_OK);
-  rc = tiledb_walk(
+  rc = tiledb_object_walk(
       ctx_, S3_TEMP_DIR.c_str(), TILEDB_POSTORDER, write_path, &walk_str);
   CHECK(rc == TILEDB_OK);
   CHECK_THAT(golden_walk, Catch::Equals(walk_str));
@@ -295,10 +477,10 @@ TEST_CASE_METHOD(WalkFx, "C API: Test walk", "[capi], [walk]") {
   golden_ls = get_golden_ls(HDFS_FULL_TEMP_DIR);
   walk_str.clear();
   ls_str.clear();
-  rc = tiledb_walk(
+  rc = tiledb_object_walk(
       ctx_, HDFS_TEMP_DIR.c_str(), TILEDB_PREORDER, write_path, &walk_str);
   CHECK(rc == TILEDB_OK);
-  rc = tiledb_walk(
+  rc = tiledb_object_walk(
       ctx_, HDFS_TEMP_DIR.c_str(), TILEDB_POSTORDER, write_path, &walk_str);
   CHECK(rc == TILEDB_OK);
   CHECK_THAT(golden_walk, Catch::Equals(walk_str));
